+++ src/main/java/eu/olli/cowmoonication/config/MooConfig.java
@@ -38,6 +38,7 @@ public class MooConfig {
     public static boolean showBestFriendNotifications;
     public static boolean showFriendNotifications;
     public static boolean showGuildNotifications;
+    public static boolean showAdvancedTooltips;
     public static String[] tabCompletableNamesCommands;
     private static String numeralSystem;
     // logs search config
@@ -127,6 +128,8 @@ public class MooConfig {
                 "showFriendNotifications", false, "Set to true to receive friends' login/logout messages, set to false hide them."), true);
         Property propShowGuildNotifications = addConfigEntry(cfg.get(Configuration.CATEGORY_CLIENT,
                 "showGuildNotifications", false, "Set to true to receive guild members' login/logout messages, set to false hide them."), true);
+        Property propShowAdvancedTooltips = addConfigEntry(cfg.get(Configuration.CATEGORY_CLIENT,
+                "showAdvancedTooltips", true, "Set to true to show advanced tooltips, set to false show default tooltips."), true);
         Property propNumeralSystem = addConfigEntry(cfg.get(Configuration.CATEGORY_CLIENT,
                 "numeralSystem", "Arabic numerals: 1, 4, 10", "Use Roman or Arabic numeral system?", new String[]{"Arabic numerals: 1, 4, 10", "Roman numerals: I, IV, X"}), true);
         Property propTabCompletableNamesCommands = addConfigEntry(cfg.get(Configuration.CATEGORY_CLIENT,
@@ -158,6 +161,7 @@ public class MooConfig {
             showBestFriendNotifications = propShowBestFriendNotifications.getBoolean();
             showFriendNotifications = propShowFriendNotifications.getBoolean();
             showGuildNotifications = propShowGuildNotifications.getBoolean();
+            showAdvancedTooltips = propShowAdvancedTooltips.getBoolean();
             numeralSystem = propNumeralSystem.getString();
             tabCompletableNamesCommands = propTabCompletableNamesCommands.getStringList();
             moo = propMoo.getString();
@@ -176,6 +180,7 @@ public class MooConfig {
         propShowBestFriendNotifications.set(showBestFriendNotifications);
         propShowFriendNotifications.set(showFriendNotifications);
         propShowGuildNotifications.set(showGuildNotifications);
+        propShowAdvancedTooltips.set(showAdvancedTooltips);
         propNumeralSystem.set(numeralSystem);
         propTabCompletableNamesCommands.set(tabCompletableNamesCommands);
         propMoo.set(moo);